move code for better presentation
